Separated declaration & implementation of Fireball.
